fix(interface/*,controllers/*,models/*): Rework API call blocks
Rework the sync process for products and carts/orders to reduce the number of commulitive API calls
made during the checkout flow. Build a process for products to be synced independently rather than
relying on the order to push what products are in its cart. Cut back on failsafe checks for things
that can be conventionally assumed. Example: stop checking for an order when creating a cart --
assume no order exists and by convention, only push to the _orders_ endpoint when an order is
completed

https://trello.com/c/dn9fgnxS/253-as-a-developer-i-should-refactor-spreechimpy-calls